doc-analyzer/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
anthropic==0.40.0
pdfplumber==0.11.4
python-docx==1.1.2
Pillow==11.0.0
pytesseract==0.3.13
pydantic==2.10.3
